+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes
 - Fix bug in magik-lint --rcfile, where java.io.File is expected but was java.io.FileInputStream
 - Fix bug where grammar accepted invalid _package identifiers, causing problems later on
 - Fix bug in DefSlottedExemplarParser: also see `sw:def_slotted_exemplar`
+- Minor clean up in `ThreadManager`
 
 
 0.5.1 (2022-01-17)